+++ rs/rosetta-api/CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.6.1] - 2022-08-25
 ### Added
 - `blockchain` command line flag that overrides the blockchain name in the network identifier.
 ### Changed
 - `NEURON_INFO` restricted data now contains followees and hotkeys.
+- `log_config.yml` now contains a specific appender for the ledger blocks synchronizer.
 
 ## [1.6.0] - 2022-05-30
 ### Fixed